split test_na3d_flash.py into separate bench and test files
